move authBanner comonent into auth folder within component
