+++ package.json
@@ -16,6 +16,7 @@
     "@reduxjs/toolkit": "^2.8.1",
     "@tailwindcss/vite": "^4.1.1",
     "axios": "^1.8.4",
+    "formik": "^2.4.6",
     "jwt-decode": "^4.0.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
@@ -27,7 +28,8 @@
     "react-swipeable": "^7.0.2",
     "recharts": "^2.15.3",
     "redux-persist": "^6.0.0",
-    "tailwindcss": "^4.1.1"
+    "tailwindcss": "^4.1.1",
+    "yup": "^1.6.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.21.0",
