+++ src/components/auth/AuthBanner.jsx
@@ -1,11 +1,11 @@
 import React, { useEffect, useState } from "react";
 import { Link } from "react-router-dom";
 import { FaArrowLeftLong } from "react-icons/fa6";
-import BannerImg01 from "../assets/images/auth-images/auth-banner-01.png";
-import BannerImg02 from "../assets/images/auth-images/auth-banner-02.png";
-import BannerImg03 from "../assets/images/auth-images/auth-banner-03.png";
-import BannerImg04 from "../assets/images/auth-images/auth-banner-04.png";
-import BannerImg05 from "../assets/images/auth-images/auth-banner-05.png";
+import BannerImg01 from "../../assets/images/auth-images/auth-banner-01.png";
+import BannerImg02 from "../../assets/images/auth-images/auth-banner-02.png";
+import BannerImg03 from "../../assets/images/auth-images/auth-banner-03.png";
+import BannerImg04 from "../../assets/images/auth-images/auth-banner-04.png";
+import BannerImg05 from "../../assets/images/auth-images/auth-banner-05.png";
 
 const banners = [
   {
