+++ src/pages/sucessful-password-rest/PasswordResetSuccess.jsx
@@ -3,7 +3,7 @@ import { Link } from "react-router-dom";
 // import SignupBanner from "../signup/SignupBanner";
 import HoralLogo from "../../assets/images/horal-logo-1.png";
 import Padlock from "../../assets/images/auth-images/padlock.gif";
-import AuthBanner from "../../components/AuthBanner";
+import AuthBanner from "../../components/auth/AuthBanner";
 
 const PasswordResetSuccess = () => {
   return (
